Prefer 'magick' command to 'convert', if available
ImageMagick version 7 provides command 'magick' (as well as legacy 'convert'), is readily available on Windows but is not, currently, readily available on Ubuntu.

On Windows, the name of ImageMagick's `convert` utility clashes with Microsoft's `C:\Windows\System32\convert.exe` utility. Using `magick`, if available, avoids the conflict.

For clarity and to be command name-agnostic, the raster option name is changed from 'convert' (the command) to 'imagemagick' (the name of the software).

For consistency with the version 7 command name, references in the code to 'convert' are changed to 'magick', throughout.