Update POV-Ray examples to avoid POV-Ray warnings
This pull request updates the POV scripts in the two POV-Ray examples to avoid the warning messages in POV-Ray 3.7 (released November 2013).

POV-Ray 3.7 requires the `version` directive.

POV-Ray 3.7 considers the `assumed_gamma` keyword to be mandatory.

POV-Ray warns when a `polygon` is not closed.

POV-Ray warns when a scaling component is omitted.