`svgBoundingPoints` for `RectangleTree` and `ImageTree`
Currently, the bounding points for a rectangle or an image are only the upper left corner and the bottom right corner (if the width and height are specified).

This pull request suggests that the other two corners of the rectangle need to be specified as bounding points (in case the rectangle is rotated by other than an integral multiple of 90 degrees).

It also adds helper functions `rectPoints` and `rectPoints'` to reduce code duplication.